Fix start/end timestamp passing.

+++ lib/net/radiotap_datalink.cpp
@@ -306,7 +306,7 @@ radiotap_datalink::parse(size_t frame_sz, const uint8_t *frame)
       int8_t junk_s8;
       uint8_t junk_u8;
       uint16_t junk_u16;
-      uint16_t start_ts, end_ts;
+      uint32_t start_ts, end_ts;
       uint64_t junk_u64;
       uint32_t bit = bitmap & i;
       switch(bit) {
